refactor(lowy): extract mode and navigation into their own modules
The morph addition pushed `tests/support/hooks.ts` past the threshold
where backend lifecycle, mode dispatch, navigation strategy, and
Cucumber plumbing all sat in one file. The volatility axes are
distinct — navigation strategy changes per mode, mode validation
changes per env contract, hook plumbing changes per Cucumber version
— so split the new code along those axes:

  - `tests/support/mode.ts`        — env validation + `Mode` type
  - `tests/support/navigation.ts`  — `openRoute`, `morphNav`,
                                     `primeMorph`, the Ema-readiness
                                     poll
  - `tests/support/hooks.ts`       — Cucumber lifecycle plumbing only

Backend launchers (`startLive`, `startStatic`) and run-root state
remain in `hooks.ts` for now — they were pre-existing and the split
of those concerns is best done when a third backend appears, not
opportunistically in this PR.

Step definitions now import navigation helpers from
`../support/navigation.ts`.
